fix: address mvcc batch publish review comments
